# config/openapi.yml
openapi: 3.0.3
info:
  title: SIONO Challenge API
  description: API requested by SIONO to myself =P
  termsOfService: http://swagger.io/terms/
  contact:
    email: simon.9864@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: /api/v1
 
paths:
  # /hello:
  #   get:
  #     description: Returns 'Hello <name>/stranger' to the caller
  #     tags: 
  #       - greeting
  #     operationId: hello
  #     parameters:
  #     - name: name
  #       required: false
  #       in: query
  #       description: The name of a caller
  #       schema:
  #         type: string
  #     responses:
  #       200:
  #         description: Success
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/HelloResponse'

  /upload/{top}:
    post:
      description: 'returns frecuency words of txt file in JSON format'
      tags:
        - Upload txt file
      parameters:
        - in: path
          name: top
          schema:
            type: integer
          required: true
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        5XX:
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - message
                properties:
                  message:
                    type: string

components:
  schemas:
    UploadResponse:
      type: object
      additionalProperties: false
      required:
        - frecuencies
      properties:
        frecuencies:
          type: array
          items:
            type: object
            properties:
              word:
                type: string
              count:
                type: integer
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message